chore: implements chart for document types
